v1 - made some great progress
Was getting worried I would have to redo a lot of my work because I
could not figure out how to make the base of each staff tip section
solid enough to connect the two. That part I resolved, it required some
bit of work and I am confident it will work when it goes to print. Also
updated the README and will be adopting this standard towards all my 3D
modeling and printing repos